replace index.ts with app.ts (setup) and server.ts (starting the server)
