--- server/src/__tests__/test_users/users.test.ts
@@ -1,3 +1,0 @@
-describe("given nothing", () => {
-  it("should pass", () => {});
-})
