+++ server/src/app.ts
@@ -1,0 +1,14 @@
+import express from "express";
+import allRoutes from "./routes/index";
+import globalMiddleware from "./middleware/globalMiddleware";
+import notFound from "./middleware/notFound";
+import errHandler from "./middleware/errHandler";
+
+const app = express();
+
+app.use(globalMiddleware);
+app.use(allRoutes);
+app.use(notFound);
+app.use(errHandler);
+
+export default app;
